delete stores warning message

+++ client/src/views/Admin.vue
@@ -88,12 +88,14 @@ export default {
   },
   methods: {
     deleteStores() {
-      Api.delete('/stores').then((res) => {
-        this.$bvModal.msgBoxOk('All Stores Are Deleted')
-      })
-        .catch(error => {
-          console.error(error)
+      if (confirm('Are you certain that you want to delete all stores? This action cannot be undone.')) {
+        Api.delete('/stores').then((res) => {
+          this.$bvModal.msgBoxOk('All Stores Are Deleted')
         })
+          .catch(error => {
+            console.error(error)
+          })
+      }
     },
     updateStore(store) {
       Api.get(`/stores/${store._id}`).then(response => {
